♻️ Move tx.go to internal/wrap
