+++ conn.go
@@ -79,7 +79,7 @@ func (c *connWrapper) Begin() (driver.Tx, error) {
 	if attr != nil {
 		lg = lg.With(*attr)
 	}
-	return wrapTx(origTx, lg), nil
+	return wrap.WrapTx(origTx, lg), nil
 }
 
 // Close implements driver.Conn.
@@ -255,5 +255,5 @@ func (c *connWithContextWrapper) BeginTx(ctx context.Context, opts driver.TxOpti
 	if attr != nil {
 		lg = lg.With(*attr)
 	}
-	return wrapTx(tx, lg), nil
+	return wrap.WrapTx(tx, lg), nil
 }
